requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
fbchat==1.0.23
idna==2.6
lxml==4.0.0
PyJWT==1.5.3
PySocks==1.6.7
python-dotenv==0.7.1
pytz==2017.2
requests==2.18.4
selenium==3.6.0
six==1.11.0
twilio==6.7.0
urllib3==1.22
